initial commit (world gen)
Initial Commit.
Generates an outer line and a set of inner lines. Generates a grid of blocks for different soil types depending on which lines it is between, and a grid of chunk objects to hold a set number of blocks. The beginning of a quadtree class for both chunk loading and for quick access to individual blocks for collisions and whatnot.